fix: added baseline auth module
